chore(infra): remove deprecated chains from warp route configs
