+++ typescript/infra/config/environments/mainnet3/agent.ts
@@ -565,27 +565,10 @@ const stagingStHyperMatchingList = chainMapMatchingList({
 // - all warp routes defined in WarpRouteIds, using addresses from the registry
 // - misc important applications not defined in the registry, e.g. merkly
 const metricAppContextsGetter = (): MetricAppContext[] => {
-  const warpContexts = Object.values(WarpRouteIds).map((warpRouteId) => {
-    let warpMatchingList = undefined;
-
-    // oUSDT has some remote routers but that don't have any limits set yet.
-    // Some people have been sending to e.g. Ink outside the UI, so to reduce alert noise
-    // we remove these from the matching list.
-    // TODO: once Ink or Worldchain have limits set, we should remove this.
-    if (warpRouteId === WarpRouteIds.oUSDT) {
-      const ousdtAddresses = getWarpAddresses(warpRouteId);
-      delete ousdtAddresses['ink'];
-      delete ousdtAddresses['worldchain'];
-      warpMatchingList = matchingList(ousdtAddresses);
-    } else {
-      warpMatchingList = warpRouteMatchingList(warpRouteId);
-    }
-
-    return {
-      name: warpRouteId,
-      matchingList: warpMatchingList,
-    };
-  });
+  const warpContexts = Object.values(WarpRouteIds).map((warpRouteId) => ({
+    name: warpRouteId,
+    matchingList: warpRouteMatchingList(warpRouteId),
+  }));
 
   return [
     ...warpContexts,
