--- typescript/infra/config/environments/mainnet3/warp/configGetters/getAncient8EthereumUSDCWarpConfig.ts
@@ -1,59 +1,0 @@
-import { ethers } from 'ethers';
-
-import {
-  ChainMap,
-  HypTokenRouterConfig,
-  OwnableConfig,
-  TokenType,
-  buildAggregationIsmConfigs,
-  defaultMultisigConfigs,
-} from '@hyperlane-xyz/sdk';
-
-import {
-  RouterConfigWithoutOwner,
-  tokens,
-} from '../../../../../src/config/warp.js';
-import { regularIcas } from '../../governance/ica/regular.js';
-import { regularSafes } from '../../governance/safe/regular.js';
-
-export const getAncient8EthereumUSDCWarpConfig = async (
-  routerConfig: ChainMap<RouterConfigWithoutOwner>,
-  abacusWorksEnvOwnerConfig: ChainMap<OwnableConfig>,
-): Promise<ChainMap<HypTokenRouterConfig>> => {
-  const ismConfig = buildAggregationIsmConfigs(
-    'ethereum',
-    ['ancient8'],
-    defaultMultisigConfigs,
-  ).ancient8;
-
-  const ethereum: HypTokenRouterConfig = {
-    ...routerConfig.ethereum,
-    ...abacusWorksEnvOwnerConfig.ethereum,
-    type: TokenType.collateral,
-    token: tokens.ethereum.USDC,
-    interchainSecurityModule: ismConfig,
-    // This hook was recovered from running the deploy script
-    // for the hook module. The hook configuration is the Ethereum
-    // default hook for the Ancient8 remote (no routing).
-    hook: '0x19b2cF952b70b217c90FC408714Fbc1acD29A6A8',
-    proxyAdmin: {
-      owner: regularSafes.ethereum,
-    },
-  };
-
-  const ancient8: HypTokenRouterConfig = {
-    ...routerConfig.ancient8,
-    ...abacusWorksEnvOwnerConfig.ancient8,
-    type: TokenType.synthetic,
-    // Uses the default ISM
-    interchainSecurityModule: ethers.constants.AddressZero,
-    proxyAdmin: {
-      owner: regularIcas.ancient8,
-    },
-  };
-
-  return {
-    ethereum,
-    ancient8,
-  };
-};
