+++ typescript/infra/config/environments/mainnet3/warp/configGetters/getRenzoREZStaging.ts
@@ -2,10 +2,7 @@ import { pick } from '@hyperlane-xyz/utils';
 
 import { getGnosisSafeBuilderStrategyConfigGenerator } from '../../../utils.js';
 
-import {
-  getRenzoWarpConfigGenerator,
-  renzoTokenPrices,
-} from './getRenzoEZETHWarpConfig.js';
+import { getRenzoWarpConfigGenerator, renzoTokenPrices } from './renzo.js';
 import {
   rezEthChainsToDeploy,
   rezEthValidators,
