+++ typescript/infra/test/warp-configs.test.ts
@@ -23,7 +23,7 @@ chai.use(chaiAsPromised);
 chai.should();
 const DEFAULT_TIMEOUT = 100000;
 
-const warpIdsToSkip: string[] = [WarpRouteIds.oUSDT, WarpRouteIds.oUSDTSTAGE];
+const warpIdsToSkip: string[] = [];
 
 async function getConfigsForBranch(branch: string) {
   return getRegistry({
